requirements.txt
flask
roboflow
numpy
pillow
opencv-python